# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "API scrapes news from various sources and provides a unified interface for users to view and manage their preferences and newsfeed also filters news based on user preferences",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/articles/details/{articleId}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get article details",
                "description": "Fetches article details by ID",
                "operationId": "b3f7dafb0ec1fb596d1be6639cb2ab1d",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "Article ID to fetch details",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article details fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/articles/{categoryId}/{sourceId}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get articles",
                "description": "Fetches articles based on category ID, source ID, and offset",
                "operationId": "3b76c6a7e4cb8f99481b3421f5b1dc57",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Category ID to filter articles",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "sourceId",
                        "in": "path",
                        "description": "Source ID to filter articles",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Articles fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "collection"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/articles/search": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Search articles",
                "description": "Searches articles based on the provided query",
                "operationId": "d9b7a795501cc77a0eb9a7fbd1635e3d",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Keyword to search articles",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Published date to search articles",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category ID to search articles",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Source ID to search articles",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Articles fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "collection"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}